Restructuring with appdelegates is complete
